refactor: API 호출 로직 movieApi.js로 분리 및 코드 정리
