+++ src/pages/Layout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import NavBar from "./Navbar";
+import NavBar from "../components/Navbar";
 
 const Layout = () => (
   <>
